updated to fit with the changed config

+++ src/main/java/com/ricardthegreat/holdmetight/mixins/PlayerMovedTooFastMixin.java
@@ -12,12 +12,12 @@ import net.minecraft.server.network.ServerGamePacketListenerImpl;
 public class PlayerMovedTooFastMixin {
     @ModifyConstant(method = "handleMovePlayer(Lnet/minecraft/network/protocol/game/ServerboundMovePlayerPacket;)V", constant = @Constant(floatValue = 100.0F))
     private float playerTooFastMaxSpeed(float speed) {
-        return HMTConfig.SERVER_CONFIG.maximumMovespeed.get();
+        return HMTConfig.SERVER_CONFIG.maximumMovespeed.get().floatValue();
     }
 
     @ModifyConstant(method = "handleMovePlayer(Lnet/minecraft/network/protocol/game/ServerboundMovePlayerPacket;)V", constant = @Constant(floatValue = 300.0F))
     private float elytraTooFastMaxSpeed(float speed) {
-        return HMTConfig.SERVER_CONFIG.maximumElytraspeed.get();
+        return HMTConfig.SERVER_CONFIG.maximumElytraspeed.get().floatValue();
     }
 
     @ModifyConstant(method = "handleMoveVehicle(Lnet/minecraft/network/protocol/game/ServerboundMoveVehiclePacket;)V", constant = @Constant(doubleValue  = 100.0F))
